Completed assignment

+++ Binomial.java
@@ -11,7 +11,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(long n, long k) { 
 		
 		if (k > n) return 0;{
 		if (k == 0 || n == 0) return 1;
@@ -21,9 +21,9 @@ public class Binomial {
 	
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
@@ -34,7 +34,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
